requirements.txt
python-decouple
requests
SQLAlchemy